Modify: Auth Usecase
